+++ CHANGELOG.md
@@ -11,6 +11,8 @@ For any change that affects end users of this package, please add an entry under
 If your change does not need a CHANGELOG entry, add the "skip changelog" label to your PR.
 
 ## Unreleased
+- Disable django instrumentation if DJANGO_SETTINGS_MODULE is not set
+  ([#549](https://github.com/aws-observability/aws-otel-python-instrumentation/pull/549))
 - Add Service and Environment dimensions to EMF metrics when Application Signals EMF export is enabled
   ([#548](https://github.com/aws-observability/aws-otel-python-instrumentation/pull/548))
 - Add Resource and CFN Attributes for Bedrock AgentCore spans